Fix the file filtering in getFileListForDirIf
The filtering for regular files did not work because the
check for the macro was mispelled.

Having fixed that, DT_REG was not defined: _XOPEN_SOURCE is
defined which causes _DEFAULT_SOURCE to not be defined
by default, see features.h. Without _DEFAULT_SOURCE, __USE_MISC
and consequently DT_REG were not defined. I therefore defined
_DEFAULT_SOURCE manually.